+++ plugins/http/src/lib.rs
@@ -36,8 +36,7 @@ pub fn init<R: Runtime>() -> TauriPlugin<R> {
         .invoke_handler(tauri::generate_handler![
             commands::fetch,
             commands::fetch_cancel,
-            commands::fetch_send,
-            commands::fetch_read_body,
+            commands::fetch_send
         ])
         .build()
 }